Fix warnings.

+++ tests/util/expr_test.cc
@@ -72,7 +72,7 @@ class TestExpr : public Expr {
 struct TestGrad {
   TestGrad *next;
   double coef;
-  double varno;
+  int varno;
 };
 
 }
